AB#88055 - criando fila separada para tratar casos de planos ativos se devem ser encerrados automaticamente pela rotina existente (bug: 87974)
