+++ src/SME.SGP.Aplicacao/Interfaces/CasosDeUso/PlanoAEE/IEncerrarPlanosAEEEstudantesInativosTratarUseCase.cs
@@ -1,0 +1,13 @@
+﻿using SME.SGP.Infra;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SME.SGP.Aplicacao.Interfaces
+{
+    public interface IEncerrarPlanosAEEEstudantesInativosTratarUseCase : IUseCase<MensagemRabbit, bool>
+    {
+    }
+}
